Allow redefine of FOTA compile time options
Usually FOTA is used as an external library to a project. There are
a few compile time options, that the user of the library can tweak
depending on the use case. Currently these can be changed only by
modifying the FOTA CMake file. Allowing user to define these options in
its CMake files simplifies the FOTA library integration.